+++ packages/salesforcedx-vscode-core/src/commands/forceSourceRetrieveMetadata/forceSourceRetrieveCmp.ts
@@ -19,7 +19,7 @@ import { RetrieveMetadataTrigger } from '.';
 import { nls } from '../../messages';
 import { SfdxPackageDirectories } from '../../sfdxProject';
 import { workspaceUtils } from '../../util';
-import { RetrieveExecutor } from '../deployRetrieveExecutor';
+import { RetrieveExecutor } from '../retrieveExecutor';
 import { SfdxCommandlet, SfdxWorkspaceChecker } from '../util';
 import { RetrieveComponentOutputGatherer } from '../util/parameterGatherers';
 import { OverwriteComponentPrompt } from '../util/postconditionCheckers';
