+++ packages/salesforcedx-vscode-core/test/vscode-integration/commands/baseDeployRetrieve.test.ts
@@ -39,11 +39,9 @@ import { SinonSpy, SinonStub, spy } from 'sinon';
 import * as vscode from 'vscode';
 import { channelService } from '../../../src/channels';
 import { BaseDeployExecutor } from '../../../src/commands';
-import {
-  DeployExecutor,
-  DeployRetrieveExecutor,
-  RetrieveExecutor
-} from '../../../src/commands/deployRetrieveExecutor';
+import { DeployExecutor } from '../../../src/commands/deployExecutor';
+import { DeployRetrieveExecutor } from '../../../src/commands/deployRetrieveExecutor';
+import { RetrieveExecutor } from '../../../src/commands/retrieveExecutor';
 import { PersistentStorageService } from '../../../src/conflict/persistentStorageService';
 import { WorkspaceContext } from '../../../src/context';
 import { getAbsoluteFilePath } from '../../../src/diagnostics';
